Refactor post form handling: consolidate form logic into usePostForm hook, streamline image upload and removal processes, and update FormData component props.
